Integración de Discord y refactorización del juego
Se ha excluido la implementación del gestor de Discord en `premake5.lua` y se ha añadido la biblioteca `discord_game_sdk`. Se ha creado una nueva sección de proyecto para `discord_game_sdk` en `premake5.lua`.

Se han añadido las funciones `InitHTP` y `UnloadHTP` en `htp.h`. Se ha eliminado código relacionado con la lógica del juego en `game.c`, `htp.c` e `intro.c`.

Se han eliminado múltiples archivos de encabezado y definiciones de variables en `main.c`, así como la función principal completa. Se ha eliminado el contenido completo de `menu.c` y `settings.c`.

Se ha añadido `discord_manager.h` con funciones para gestionar la presencia de Discord. Se ha añadido `game.cpp` con la implementación de `DrawGame`.

Se han añadido inclusiones de bibliotecas necesarias en varios archivos. Se han implementado funciones para gestionar la pantalla "How to Play", la introducción del juego, el menú principal y la configuración del juego.

Se ha añadido la inicialización de la ventana del juego y la configuración de Discord Rich Presence en `main.cpp`. Se ha añadido la carga y descarga de texturas y sonidos del juego.

Se han añadido varias inclusiones de cabeceras en `discord_manager.cpp`. Se ha introducido un espacio de nombres `discord_sdk` con funciones para inicializar, actualizar y cerrar la presencia de Discord.

+++ src/intro.cpp
@@ -1,8 +1,9 @@
 #include "intro.h"
+#include <string>
+
 #define MAX_SOUNDS 10
 
-
-void FadeInOutLogo(Texture2D texture, const char* companyText, int fadeDuration, int screenWidth, int screenHeight) 
+void FadeInOutLogo(Texture2D texture, const char* companyText, int fadeDuration, int screenWidth, int screenHeight)
 {
     int fadeInDuration = fadeDuration / 2;
     int fadeOutDuration = fadeDuration / 2;
@@ -15,9 +16,9 @@ void FadeInOutLogo(Texture2D texture, const char* companyText, int fadeDuration,
     int maxLogoHeight = screenHeight - companyTextSize - marginBetweenLogoAndText;
     float logoScale = 1.0f;
 
-    if (texture.height > maxLogoHeight) 
+    if (texture.height > maxLogoHeight)
     {
-        logoScale = (float)maxLogoHeight / texture.height;
+        logoScale = static_cast<float>(maxLogoHeight) / texture.height;
     }
 
     int logoDrawWidth = texture.width * logoScale;
@@ -33,27 +34,27 @@ void FadeInOutLogo(Texture2D texture, const char* companyText, int fadeDuration,
 
     int framesCounter = 0;
 
-    while (framesCounter < totalDuration) 
+    while (framesCounter < totalDuration)
     {
         float alpha;
 
-        if (framesCounter < fadeInDuration) 
+        if (framesCounter < fadeInDuration)
         {
-            alpha = (float)framesCounter / fadeInDuration;
+            alpha = static_cast<float>(framesCounter) / fadeInDuration;
         }
-        else if (framesCounter < fadeInDuration + pauseDuration) 
+        else if (framesCounter < fadeInDuration + pauseDuration)
         {
             alpha = 1.0f;
         }
-        else 
+        else
         {
-            alpha = 1.0f - (float)(framesCounter - fadeInDuration - pauseDuration) / fadeOutDuration;
+            alpha = 1.0f - static_cast<float>(framesCounter - fadeInDuration - pauseDuration) / fadeOutDuration;
         }
 
         BeginDrawing();
         ClearBackground(BLACK);
         if (texture.id != 0) {
-            DrawTextureEx(texture, (Vector2) { logoX, logoY }, 0.0f, logoScale, Fade(WHITE, alpha));
+            DrawTextureEx(texture, Vector2{ static_cast<float>(logoX), static_cast<float>(logoY) }, 0.0f, logoScale, Fade(WHITE, alpha));
         }
         DrawText(companyText, companyTextX, companyTextY, companyTextSize, Fade(WHITE, alpha));
         EndDrawing();
@@ -62,7 +63,7 @@ void FadeInOutLogo(Texture2D texture, const char* companyText, int fadeDuration,
     }
 }
 
-void FadeInOutText(const char* text, int textSize, int fadeDuration, int screenWidth, int screenHeight) 
+void FadeInOutText(const char* text, int textSize, int fadeDuration, int screenWidth, int screenHeight)
 {
     int fadeInDuration = fadeDuration / 2;
     int fadeOutDuration = fadeDuration / 2;
@@ -74,20 +75,20 @@ void FadeInOutText(const char* text, int textSize, int fadeDuration, int screenW
 
     int framesCounter = 0;
 
-    while (framesCounter < totalDuration) 
+    while (framesCounter < totalDuration)
     {
         float alpha;
-        if (framesCounter < fadeInDuration) 
+        if (framesCounter < fadeInDuration)
         {
-            alpha = (float)framesCounter / fadeInDuration;
+            alpha = static_cast<float>(framesCounter) / fadeInDuration;
         }
-        else if (framesCounter < fadeInDuration + pauseDuration) 
+        else if (framesCounter < fadeInDuration + pauseDuration)
         {
             alpha = 1.0f;
         }
-        else 
+        else
         {
-            alpha = 1.0f - (float)(framesCounter - fadeInDuration - pauseDuration) / fadeOutDuration;
+            alpha = 1.0f - static_cast<float>(framesCounter - fadeInDuration - pauseDuration) / fadeOutDuration;
         }
 
         BeginDrawing();
@@ -99,7 +100,7 @@ void FadeInOutText(const char* text, int textSize, int fadeDuration, int screenW
     }
 }
 
-void ShowIntro(int screenWidth, int screenHeight, Sound fxStart) 
+void ShowIntro(int screenWidth, int screenHeight, Sound fxStart)
 {
     Texture2D icon = LoadTexture("icon.png");
     Texture2D JotPKLogo = LoadTexture("JotPK.png");
@@ -130,11 +131,11 @@ void ShowIntro(int screenWidth, int screenHeight, Sound fxStart)
     bool showPressStart = true;
     int framesCounter = 0;
 
-    while (!IsKeyPressed(KEY_SPACE)) 
+    while (!IsKeyPressed(KEY_SPACE))
     {
         framesCounter++;
 
-        if (framesCounter % 60 < 30) 
+        if (framesCounter % 60 < 30)
         {
             showPressStart = true;
         }
@@ -144,9 +145,9 @@ void ShowIntro(int screenWidth, int screenHeight, Sound fxStart)
 
         BeginDrawing();
         ClearBackground(BLACK);
-        DrawTextureEx(JotPKLogo, (Vector2) { logoX, logoY }, 0.0f, (float)logoWidth / JotPKLogo.width, WHITE);
+        DrawTextureEx(JotPKLogo, Vector2{ static_cast<float>(logoX), static_cast<float>(logoY) }, 0.0f, static_cast<float>(logoWidth) / JotPKLogo.width, WHITE);
 
-        if (showPressStart) 
+        if (showPressStart)
         {
             DrawTexture(pressStart, pressStartX, pressStartY, WHITE);
         }
@@ -162,12 +163,12 @@ void ShowIntro(int screenWidth, int screenHeight, Sound fxStart)
     // Parpadeo rápido de "press_start" al presionar espacio
 
     for (int i = 0; i < 10; i++) // Aumentar el número de iteraciones para que dure más tiempo
-    { 
+    {
         BeginDrawing();
         ClearBackground(BLACK);
-        DrawTextureEx(JotPKLogo, (Vector2) { logoX, logoY }, 0.0f, (float)logoWidth / JotPKLogo.width, WHITE);
+        DrawTextureEx(JotPKLogo, Vector2{ static_cast<float>(logoX), static_cast<float>(logoY) }, 0.0f, static_cast<float>(logoWidth) / JotPKLogo.width, WHITE);
 
-        if (i % 2 == 0) 
+        if (i % 2 == 0)
         {
             DrawTexture(pressStart, pressStartX, pressStartY, WHITE);
         }
@@ -178,11 +179,11 @@ void ShowIntro(int screenWidth, int screenHeight, Sound fxStart)
 
     // Pausa en negro durante 1.5 segundos con el logo de JotPK
 
-    for (int i = 0; i < 15; i++) 
+    for (int i = 0; i < 15; i++)
     {
         BeginDrawing();
         ClearBackground(BLACK);
-        DrawTextureEx(JotPKLogo, (Vector2) { logoX, logoY }, 0.0f, (float)logoWidth / JotPKLogo.width, WHITE);
+        DrawTextureEx(JotPKLogo, Vector2{ static_cast<float>(logoX), static_cast<float>(logoY) }, 0.0f, static_cast<float>(logoWidth) / JotPKLogo.width, WHITE);
         EndDrawing();
         WaitTime(0.1); // Esperar 0.1 segundos
     }
